requirements.txt
Flask==2.3.2
boto3==1.28.57
botocore==1.31.57
Pillow==10.1.0
requests==2.31.0